Integer Id zu Long Id geändert

+++ src/main/java/com/example/demo/ListItemRepository.java
@@ -2,6 +2,6 @@ package com.example.demo;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface ListItemRepository extends CrudRepository<ListItem, Integer> {}
+public interface ListItemRepository extends CrudRepository<ListItem, Long> {}
